pyproject.toml
[project]
name = "imc-prosperity-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=4.1.0",
    "dash-bootstrap-components>=2.0.4",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pathlib>=1.0.1",
    "plotly>=6.7.0",
    "polars==1.2.1",
    "prosperity4btest>=0.4.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "statsmodels>=0.14.6",
]
